+++ src/App/Identity/identity/Color/swedbankpay-colors.js
@@ -1,208 +1,358 @@
-export const brandColors = [
+export const primaryBrandColors = [
 	{
-		title: "Yellow",
+		title: "Primary",
 		subTitle: "brand-primary",
 		hex: "#FDC129",
-		contrastBlackApproved: true,
 	},
 	{
-		title: "Yellow light",
+		title: "Primary light",
 		subTitle: "brand-primary-light",
 		hex: "#FBDD91",
-		contrastBlackApproved: true,
 	},
 	{
-		title: "Yellow lighter",
+		title: "Primary lighter",
 		subTitle: "brand-primary-lighter",
 		hex: "#FFF3D5",
-		contrastBlackApproved: true,
 	},
 ];
 
-export const complementaryColors = [
+export const accentBrandColors = [
+	{
+		title: "Brown",
+		subTitle: "brand-brown",
+		hex: "#2F2424",
+	},
 	{
 		title: "Orange",
 		subTitle: "brand-orange",
 		hex: "#EE7023",
-		contrastWhiteAlert: true,
-		contrastBlackApproved: true,
-	},
-	{
-		title: "Turquoise",
-		subTitle: "brand-turquoise",
-		hex: "#31A3AE",
-		contrastWhiteAlert: true,
-		contrastBlackApproved: true,
-	},
-	{
-		title: "Turquoise light",
-		subTitle: "brand-turquoise-light",
-		hex: "#8ACDC3",
-		contrastBlackApproved: true,
 	},
 	{
 		title: "Pink",
 		subTitle: "brand-pink",
 		hex: "#EFB7B6",
-		contrastBlackApproved: true,
+	},
+	{
+		title: "Turquoise",
+		subTitle: "brand-turquoise",
+		hex: "#31A3AE",
+	},
+	{
+		title: "Turquoise light",
+		subTitle: "brand-turquoise-light",
+		hex: "#8ACDC3",
 	},
 ];
 
 export const textColors = [
 	{
-		title: "Brown solid",
-		subTitle: "brown-solid",
+		title: "Text default",
+		subTitle: "text-default",
 		hex: "#2F2424",
-		contrastWhiteApproved: true,
-		noBlackExample: true,
+		textCategory: "default-mode",
 	},
 	{
-		title: "Brown medium",
+		title: "Text muted",
 		subTitle: "text-muted",
 		hex: "#72605E",
-		contrastWhiteApproved: true,
+		textCategory: "default-mode",
 	},
 	{
-		title: "Turquoise dark",
-		subTitle: "text-link",
-		hex: "#257886",
-		contrastWhiteApproved: true,
-	},
-	{
-		title: "White",
+		title: "Text White",
 		subTitle: "text-white",
 		hex: "#FFFFFF",
-		border: "has-border",
-		contrastBlackApproved: true,
-		noWhiteExample: true,
+		hasBorderOverWhiteBG: true,
+		textCategory: "dark-backgrounds",
+	},
+	{
+		title: "Text Muted White",
+		subTitle: "text-muted-light",
+		hex: "##A38B80",
+		textCategory: "dark-backgrounds",
+	},
+	{
+		title: "Text Disabled",
+		subTitle: "text-disabled",
+		hex: "#D4C4BC",
+		textCategory: "disabled",
 	},
 ];
 
 export const backgroundColors = [
 	{
-		title: "White",
+		title: "BG Orange",
+		subTitle: "bg-orange",
+		hex: "#FFF1E4",
+	},
+	{
+		title: "BG Brown",
+		subTitle: "bg-brown",
+		hex: "#EBE7E2",
+	},
+	{
+		title: "BG White",
 		subTitle: "bg-white",
 		hex: "#FFFFFF",
-		border: "has-border",
-		contrastBlackApproved: true,
-		noWhiteExample: true,
+		hasBorderOverWhiteBG: true,
 	},
 	{
-		title: "Apricot",
-		subTitle: "bg-primary",
-		hex: "#FBF2EA",
-		contrastBlackApproved: true,
-	},
-	{
-		title: "Turquoise lighter",
-		subTitle: "bg-secondary",
+		title: "BG Turquoise",
+		subTitle: "bg-turquoise",
 		hex: "#EBF8F2",
-		contrastBlackApproved: true,
 	},
 	{
-		title: "Gray",
-		subTitle: "bg-tertiary",
-		hex: "#EBE7E2",
-		contrastBlackApproved: true,
-	},
-	{
-		title: "Gray light",
-		subTitle: "bg-quaternary",
+		title: "BG Gray",
+		subTitle: "bg-grey",
 		hex: "#F9F8F6",
-		contrastBlackApproved: true,
-		border: "has-border",
+		hasBorderOverWhiteBG: true,
 	},
 ];
 
 export const systemColors = [
 	{
-		title: "Error dark",
-		subTitle: "system-error-dark",
-		hex: "#8c0d14",
-		contrastWhiteApproved: true,
+		title: "Success",
+		subTitle: "system-success",
+		hex: "#4D9C0F",
 	},
 	{
-		title: "Error",
-		subTitle: "system-error",
-		hex: "#B5111A",
-		contrastWhiteApproved: true,
-	},
-	{
-		title: "Error light",
-		subTitle: "system-error-light",
-		hex: "#FCEEE9",
-		contrastBlackApproved: true,
+		title: "Success light",
+		subTitle: "system-success-light",
+		hex: "#EFF8D2",
 	},
 	{
 		title: "Warning",
 		subTitle: "system-warning",
 		hex: "#EE7023",
-		contrastWhiteApproved: true,
-		contrastBlackApproved: true,
 	},
 	{
 		title: "Warning light",
 		subTitle: "system-warning-light",
-		hex: "#FEF2E2",
-		contrastBlackApproved: true,
-	},
-	{
-		title: "Success",
-		subTitle: "system-success",
-		hex: "#3C7213",
-		contrastWhiteApproved: true,
-	},
-	{
-		title: "Success light",
-		subTitle: "system-success-light",
-		hex: "#F2F7EB",
-		contrastBlackApproved: true,
+		hex: "#FBEADB",
 	},
 	{
 		title: "Info",
 		subTitle: "system-info",
-		hex: "#3961A7",
-		contrastWhiteApproved: true,
+		hex: "#335BB1",
 	},
 	{
 		title: "Info light",
 		subTitle: "system-info-light",
-		hex: "#EBF1FA",
-		contrastBlackApproved: true,
+		hex: "#E0EAF9",
+	},
+	{
+		title: "Error",
+		subTitle: "system-error",
+		hex: "#D82E2A",
+	},
+	{
+		title: "Error light",
+		subTitle: "system-error-light",
+		hex: "#FFE9E9",
 	},
 ];
 
 export const functionalColors = [
 	{
-		title: "Brown solid",
-		subTitle: "brown-solid",
 		hex: "#2F2424",
-		contrastWhiteApproved: true,
-		noBlackExample: true,
+		baseTokenTitle: "Base 1",
+		baseTokenSubTitle: "base-1",
+		specificTokenTitle: "Text Default",
+		specificTokenSubTitle: "text-default",
+		description: "The default text and icon color on light backgrounds.",
 	},
 	{
-		title: "Brown medium",
-		subTitle: "brown-medium",
 		hex: "#72605E",
-		contrastWhiteApproved: true,
+		baseTokenTitle: "Base 2",
+		baseTokenSubTitle: "base-2",
+		specificTokenTitle: "Text Muted",
+		specificTokenSubTitle: "text-muted",
+		description:
+			"For secondary or less important text that shouldn’t be the primary focus.",
 	},
 	{
-		title: "Brown soft",
-		subTitle: "brown-soft",
 		hex: "#A38B80",
-		contrastWhiteAlert: true,
-		contrastBlackApproved: true,
+		baseTokenTitle: "Base 3",
+		baseTokenSubTitle: "base-3",
+		specificTokenTitle: "Text Muted light",
+		specificTokenSubTitle: "text-muted-light",
+		description: "The secondary text color for dark backgrounds.",
 	},
 	{
-		title: "Brown light",
-		subTitle: "brown-light",
 		hex: "#D4C4BC",
-		contrastBlackApproved: true,
+		baseTokenTitle: "Base 4",
+		baseTokenSubTitle: "base-4",
+		specificTokenTitle: "Text Disabled",
+		specificTokenSubTitle: "text-disabled",
+		description:
+			"Dividers and other low contrast UI elements. It is also used for disabled text on white background but also on buttons.",
 	},
 	{
-		title: "Gray",
-		subTitle: "bg-tertiary",
 		hex: "#EBE7E2",
-		contrastBlackApproved: true,
+		baseTokenTitle: "Base 5",
+		baseTokenSubTitle: "base-5",
+		specificTokenTitle: "BG Brown",
+		specificTokenSubTitle: "bg-brown",
+		description:
+			"Background color emphasized surfaces such as buttons. This is also the bg color of disabled main action buttons.",
+	},
+	{
+		hex: "#F9F8F6",
+		hasBorderOverWhiteBG: true,
+		baseTokenTitle: "Base 6",
+		baseTokenSubTitle: "base-6",
+		specificTokenTitle: "BG Grey",
+		specificTokenSubTitle: "bg-grey",
+		description: "Default background color.",
+	},
+	{
+		hex: "#FFFFFF",
+		hasBorderOverWhiteBG: true,
+		baseTokenTitle: "White",
+		baseTokenSubTitle: "white",
+		amountSpecificToken: 2,
+		specificTokenTitle: "Text White",
+		specificTokenSubTitle: "text-white",
+		specificTokenTitle2: "BG White",
+		specificTokenSubTitle2: "bg-white",
+		description:
+			"Used for emphasized surfaces on colors background, like search fields. The default text and icon color for dark backgrounds.",
+	},
+];
+
+export const colorPool = [
+	{
+		title: "Primary",
+		subTitle: "brand-yellow",
+		hex: "#FDC129",
+	},
+	{
+		title: "Yellow light",
+		subTitle: "brand-yellow-light",
+		hex: "#FCE4A1",
+	},
+	{
+		title: "Yellow lighter",
+		subTitle: "brand-yellow-lighter",
+		hex: "#FFF3D5",
+	},
+	{
+		title: "Turquoise",
+		subTitle: "brand-turquoise",
+		hex: "#31A3AE",
+	},
+	{
+		title: "Turquoise light",
+		subTitle: "brand-turquoise-light",
+		hex: "#8ACDC3",
+	},
+	{
+		title: "Turquoise lighter",
+		subTitle: "brand-turquoise-lighter",
+		hex: "#EBF8F2",
+	},
+	{
+		title: "Purple",
+		subTitle: "brand-purple",
+		hex: "#642f93",
+	},
+	{
+		title: "Purple light",
+		subTitle: "brand-purple-light",
+		hex: "#C58AF9",
+	},
+	{
+		title: "Purple lighter",
+		subTitle: "brand-purple-lighter",
+		hex: "#F1E0FF",
+	},
+	{
+		title: "Orange",
+		subTitle: "brand-orange",
+		hex: "#EE7023",
+	},
+	{
+		title: "Orange light",
+		subTitle: "brand-orange-light",
+		hex: "#FFBA69",
+	},
+	{
+		title: "Orange lighter",
+		subTitle: "brand-orange-lighter",
+		hex: "#FBEADB",
+	},
+	{
+		title: "Green",
+		subTitle: "brand-green",
+		hex: "#497125",
+	},
+	{
+		title: "Green light",
+		subTitle: "brand-green-light",
+		hex: "#70B63A",
+	},
+	{
+		title: "Green lighter",
+		subTitle: "brand-green-lighter",
+		hex: "#E7F8D2",
+	},
+	{
+		title: "Mahogany",
+		subTitle: "brand-mahogany",
+		hex: "#4C2D2C",
+	},
+	{
+		title: "Mahogany light",
+		subTitle: "brand-mahogany-light",
+		hex: "#965E5C",
+	},
+	{
+		title: "Mahogany lighter",
+		subTitle: "brand-mahogany-lighter",
+		hex: "#EEC4C4",
+	},
+	{
+		title: "Red",
+		subTitle: "brand-red",
+		hex: "#A62825",
+	},
+	{
+		title: "Red light",
+		subTitle: "brand-red-light",
+		hex: "#EFB7B6",
+	},
+	{
+		title: "Red lighter",
+		subTitle: "brand-red-lighter",
+		hex: "#FFE9E9",
+	},
+	{
+		title: "Blue",
+		subTitle: "brand-blue",
+		hex: "#335BB1",
+	},
+	{
+		title: "Blue light",
+		subTitle: "brand-blue-light",
+		hex: "#65A0E4",
+	},
+	{
+		title: "Blue lighter",
+		subTitle: "brand-blue-lighter",
+		hex: "#E0EAF9",
+	},
+	{
+		title: "Fuchsia",
+		subTitle: "brand-fuchsia",
+		hex: "#B9297F",
+	},
+	{
+		title: "Fuchsia light",
+		subTitle: "brand-fuchsia-light",
+		hex: "#FBBBE1",
+	},
+	{
+		title: "Fuchsia lighter",
+		subTitle: "brand-fuchsia-lighter",
+		hex: "#FFEEFC",
 	},
 ];