+++ RELEASE_NOTES.md
@@ -6,6 +6,10 @@
 
 ## Component changes
 
+- Colors utilities:
+  - removed the possibility to combine text-ciolor utility classes WITH `.text-muted` to decrease their opacity
+  - `.text-muted` class still exist, but it will always mean a full opacity color
+
 ### Non breaking UI changes
 
 - add possibility to disable rotation of the dropdown toggle icon on active state
@@ -18,3 +22,4 @@
 ## Design System website
 
 - expand components preview containers to full screen
+- Update the Colors page for SwedbankPay side